# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Pit-to-Port Weighbridge Reconciliation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ReconcileInput {
  /** Mine weighbridge slip (PDF or image) */
  mine_slip: Blob;
  /** Port siding receipt (PDF or image) */
  port_slip: Blob;
}

export interface ReconciliationLog {
  id: number;
  /** @nullable */
  consignment_note?: string | null;
  /** @nullable */
  port_reference?: string | null;
  truck_reg: string;
  mine_net_weight: number;
  port_net_weight: number;
  variance: number;
  /** @nullable */
  transit_hours?: number | null;
  status: string;
  /** @nullable */
  raw_mine_json?: string | null;
  /** @nullable */
  raw_port_json?: string | null;
  /** @nullable */
  corrected_by?: string | null;
  created_at: string;
}

export interface ReconciliationUpdate {
  /** @nullable */
  consignment_note?: string | null;
  /** @nullable */
  port_reference?: string | null;
  truck_reg: string;
  mine_net_weight: number;
  port_net_weight: number;
}

export interface ReconciliationStats {
  total: number;
  reconciled: number;
  warnings: number;
  critical: number;
  avg_variance: number;
  /** @nullable */
  avg_transit_hours: number | null;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  operator: 'operator',
  admin: 'admin',
} as const;

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  role?: AuthUserRole;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type ListReconciliationsParams = {
/**
 * Filter by truck registration (case-insensitive)
 */
search?: string;
/**
 * Filter by status value
 */
status?: string;
/**
 * Filter records on or after this date (YYYY-MM-DD)
 */
date_from?: string;
/**
 * Filter records on or before this date (YYYY-MM-DD)
 */
date_to?: string;
};

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};