# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pit-to-Port Weighbridge Reconciliation API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reconciliation
    description: Weighbridge reconciliation operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reconcile:
    post:
      operationId: reconcileDocuments
      tags: [reconciliation]
      summary: Process and reconcile mine + port documents
      description: Uploads two documents (mine slip and port receipt), extracts data via Gemini, runs reconciliation logic, and persists the result.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ReconcileInput"
      responses:
        "201":
          description: Reconciliation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationLog"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server or Gemini error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reconciliations:
    get:
      operationId: listReconciliations
      tags: [reconciliation]
      summary: List all reconciliation logs
      description: Returns all reconciliation log entries ordered by most recent first.
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Filter by truck registration (case-insensitive)
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by status value
        - name: date_from
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filter records on or after this date (YYYY-MM-DD)
        - name: date_to
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Filter records on or before this date (YYYY-MM-DD)
      responses:
        "200":
          description: List of reconciliation logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReconciliationLog"

  /reconciliations/stats:
    get:
      operationId: getReconciliationStats
      tags: [reconciliation]
      summary: Aggregate statistics
      description: Returns summary counts and averages across all reconciliation logs.
      responses:
        "200":
          description: Aggregated stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationStats"

  /reconciliations/{id}:
    get:
      operationId: getReconciliation
      tags: [reconciliation]
      summary: Get single reconciliation log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reconciliation log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationLog"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateReconciliation
      tags: [reconciliation]
      summary: Manually correct a reconciliation log
      description: Human-in-the-loop override to correct extracted values and recalculate status.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconciliationUpdate"
      responses:
        "200":
          description: Updated reconciliation log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationLog"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteReconciliation
      tags: [reconciliation]
      summary: Delete a reconciliation log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'

components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    ReconcileInput:
      type: object
      required: [mine_slip, port_slip]
      properties:
        mine_slip:
          type: string
          format: binary
          description: Mine weighbridge slip (PDF or image)
        port_slip:
          type: string
          format: binary
          description: Port siding receipt (PDF or image)

    ReconciliationLog:
      type: object
      required:
        - id
        - truck_reg
        - mine_net_weight
        - port_net_weight
        - variance
        - status
        - created_at
      properties:
        id:
          type: integer
        consignment_note:
          type: ["string", "null"]
        port_reference:
          type: ["string", "null"]
        truck_reg:
          type: string
        mine_net_weight:
          type: number
        port_net_weight:
          type: number
        variance:
          type: number
        transit_hours:
          type: ["number", "null"]
        status:
          type: string
        raw_mine_json:
          type: ["string", "null"]
        raw_port_json:
          type: ["string", "null"]
        corrected_by:
          type: ["string", "null"]
        created_at:
          type: string
          format: date-time

    ReconciliationUpdate:
      type: object
      required: [truck_reg, mine_net_weight, port_net_weight]
      properties:
        consignment_note:
          type: ["string", "null"]
        port_reference:
          type: ["string", "null"]
        truck_reg:
          type: string
        mine_net_weight:
          type: number
        port_net_weight:
          type: number

    ReconciliationStats:
      type: object
      required:
        - total
        - reconciled
        - warnings
        - critical
        - avg_variance
        - avg_transit_hours
      properties:
        total:
          type: integer
        reconciled:
          type: integer
        warnings:
          type: integer
        critical:
          type: integer
        avg_variance:
          type: number
        avg_transit_hours:
          type: ["number", "null"]

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
        role:
          type: string
          enum: [operator, admin]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string